+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6
 
 * Add `unionTagName` option to customize the `"Case"` tag for unions.
+* Add `JsonUnionEncoding.FSharpLuLike`, which is equivalent to `ExternalTag ||| BareFieldlessTags`.
 
 ## 0.5